requirements.txt
 python-binance
 ta
 pandas
 mpl_finance